requirements.txt
asgiref==3.3.1
dj-database-url==0.5.0
Django==3.1.3
django-model-utils==4.0.0
psycopg2==2.8.6
pytz==2020.4
sqlparse==0.4.1
